XlsFormat - исправление конфликтов сборки под QT
git-svn-id: svn://fileserver/activex/AVS/Sources/TeamlabOffice/trunk/ServerComponents@68357 954022d7-b5bf-4e40-9824-e11837661b57

+++ ASCOfficeXlsFile2/source/XlsFormat/Logic/Biff_records/UserSViewBegin.cpp
@@ -143,9 +143,9 @@ void UserSViewBegin::readFields(CFRecord& record)
 	fPageLayoutView = GETBIT(flags, 11);
 	fRuler			= GETBIT(flags, 13);
 
-	record >> operNumX >> operNumY >> colRPane >> rwBPane;
+    record >> operNumX >> operNumY >> colRPane >> rwBPane;
 
-	pane_top_left_cell = CellRef(fSplitH ? rwBPane : 0 , fSplitV ? colRPane : 0, true, true).toString();
+    pane_top_left_cell = CellRef(fSplitH ? (int)rwBPane : 0 , fSplitV ? (int)colRPane : 0, true, true).toString();
 }
 
 } // namespace XLS
